requirements.txt
asgiref==3.4.1
Django==3.2.7
django-model-utils==4.1.1
django-notifications-hq==1.6.0
gunicorn==20.1.0
jsonfield==3.1.0
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.1
swapper==1.1.2.post1
whitenoise==5.3.0
